requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
strands-agents[ollama]>=0.1.0
strands-agents-tools>=0.1.0
python-dotenv>=1.0.1
pydantic>=2.8.0
